Refactor OrderPayment -> OrderSummaryCard
