+++ src/orders/components/OrderSummaryCard/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./OrderSummaryCard";
+export * from "./OrderSummaryCard";
